server/requirements.txt
alembic==1.14.1
appdata==2.2.1
apturl==0.5.2
bcrypt==3.1.7
blinker==1.8.2
Brlapi==0.7.0
certifi==2019.11.28
chardet==3.0.4
click==8.1.8
colorama==0.4.3
colored==1.4.4
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distlib==0.3.9
distro==1.4.0
distro-info==0.23+ubuntu1.1
duplicity==0.8.12.0
entrypoints==0.3
Faker==33.1.0
fasteners==0.14.1
filelock==3.16.1
Flask==3.0.3
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
flask-marshmallow==1.2.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
future==0.18.2
greenlet==3.1.1
gunicorn==23.0.0
httplib2==0.14.0
idna==2.8
importlib_metadata==8.5.0
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.5
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.3.8
MarkupSafe==2.1.5
marshmallow==3.22.0
marshmallow-sqlalchemy==1.1.1
monotonic==1.5
numpy==2.2.1
oauthlib==3.1.0
olefile==0.46
packaging==24.2
pandas==2.2.3
paramiko==2.6.0
pexpect==4.6.0
Pillow==7.0.0
pipenv==2024.4.0
platformdirs==4.3.6
protobuf==3.6.1
psycopg2-binary==2.9.10
pycairo==1.16.2
pycups==1.9.73
PyGObject==3.36.0
PyJWT==2.9.0
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyRFC3339==1.1
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.9.0.post0
python-debian==0.1.36+ubuntu1.1
python-dotenv==1.0.1
pytz==2024.2
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
SQLAlchemy==2.0.37
systemd-python==234
to-requirements.txt==2.0.11
typing_extensions==4.12.2
tzdata==2024.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
uuid==1.30
virtualenv==20.28.0
wadllib==1.3.3
Werkzeug==3.0.6
xkit==0.0.0
zipp==3.20.2
